+++ guillotina_rediscache/interfaces.py
@@ -1,7 +1,4 @@
-try:
-    from guillotina.async_util import IAsyncUtility
-except ImportError:
-    from guillotina.async import IAsyncUtility
+from guillotina.async_util import IAsyncUtility
 
 
 CACHE_PREFIX = 'gcache2-'